requirements.txt
tflite-runtime
numpy==1.26.4
python-chess
Flask
Flask-Cors
requests